Add minimally functional behavior sending code to the symbiot repo
The sketch named 'LB_sender' has the main functionality of sending
a light behavior over the rf24 radio, as well as having radios listen
for incoming light behaviors and actuate them. Currently, everything
is hardcoded: the transmitted LB, the radio addresses (which need
to be addressed in the long run), and the LB itself (which is currently
a number).

Despite this, the two nodes are able to transmit to one another
without entering an unsafe state and exhibiting undefined behavior.
The next steps are to change the transmitted data to an actual LB
and having the receiving node actuate.