[UNTESTED] Implement BeanSockets and add skeletons for Bean and Behavior
BeanSockets takes the working lightscript-sending capability in the experimental
LB_sender.ino sketch and modularizes it into the BeanSocket object. At the moment
this is a major part of the application layer.

Currently, BlinkMs and their associated methods are hard-coded as the only
possible actuator, while in the future, once I figure out how to deal with
and wrap multiple actuators' methods, I can generalize the actuator field.

Note that the BeanSocket class should be taken for a test drive. Next, I will
implement the Bean and Behavior classes.