Change fields from objects to pointers and implement Behavior.
Previously, Beansocket held copies of its RF24 and other objects within the class.
I changed this to pointers and because of this, was able to simplify the code
for Beansocket's implementation. Though, Behavior still holds a copy of the array
of lightscript entries because Behavior is intended to be a wrapper for the
array.